Created new class and object in main
I created a Train class and made three instance variables as well as a method to write a string dictating the noise a train makes. I then used that class to create a train object in the main method, and used the method from the train class to print the noise a train makes.